git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@13352 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/TextChunkManager.cxx
@@ -38,7 +38,7 @@ TextChunk::TextChunk(const std::string& _fileName, int _maxLines, int _maxLineLe
 {
   fileName = _fileName;
   maxLines = _maxLines;
-  maxLineLength = _maxLineLength < 0 ? std::string::npos : _maxLineLength;
+  maxLineLength = _maxLineLength < 0 ? (int)std::string::npos : _maxLineLength;
   theVector = parse();
   return;
 }
@@ -76,7 +76,7 @@ StringVector TextChunk::parse()
     }
 
     // truncate long lines
-    if (line.size() > maxLineLength) {
+    if ((int)line.size() > maxLineLength) {
       line.erase(maxLineLength);
       long_lines_encountered++;
     }
